refactor: PR 반영(3) - usecase 오타 수정 - 테스트를 위한 ReadService 추가
