+++ dailyge-api/src/main/java/project/dailyge/app/core/anniversary/application/AnniversaryReadService.java
@@ -1,0 +1,7 @@
+package project.dailyge.app.core.anniversary.application;
+
+import project.dailyge.entity.anniversary.AnniversaryJpaEntity;
+
+public interface AnniversaryReadService {
+    AnniversaryJpaEntity findById(Long anniversaryId);
+}
