+++ dailyge-api/src/main/java/project/dailyge/app/core/anniversary/application/usecase/AnniversaryReadUseCase.java
@@ -1,0 +1,22 @@
+package project.dailyge.app.core.anniversary.application.usecase;
+
+import lombok.RequiredArgsConstructor;
+import project.dailyge.app.common.annotation.ApplicationLayer;
+import project.dailyge.app.core.anniversary.application.AnniversaryReadService;
+import static project.dailyge.app.core.anniversary.exception.AnniversaryCodeAndMessage.ANNIVERSARY_NOT_FOUND;
+import project.dailyge.app.core.anniversary.exception.AnniversaryTypeException;
+import project.dailyge.entity.anniversary.AnniversaryEntityReadRepository;
+import project.dailyge.entity.anniversary.AnniversaryJpaEntity;
+
+@RequiredArgsConstructor
+@ApplicationLayer(value = "AnniversaryReadUseCase")
+class AnniversaryReadUseCase implements AnniversaryReadService {
+
+    private final AnniversaryEntityReadRepository anniversaryEntityReadRepository;
+
+    @Override
+    public AnniversaryJpaEntity findById(final Long anniversaryId) {
+        return anniversaryEntityReadRepository.findById(anniversaryId)
+            .orElseThrow(() -> AnniversaryTypeException.from(ANNIVERSARY_NOT_FOUND));
+    }
+}
